Move remaining state to atoms
