+++ src/index.tsx
@@ -7,7 +7,6 @@ import reportWebVitals from "reportWebVitals";
 import ThemeProvider from "components/Theme/ThemeProvider";
 import GlobalStyles from "components/Theme/GlobalStyles";
 import AuthProvider from "components/Auth/AuthProvider";
-import GameProvider from "components/Game/GameProvider";
 import PageRouter from "pages/PageRouter";
 
 const container = document.getElementById("root");
@@ -19,10 +18,8 @@ if (container) {
     <React.StrictMode>
       <AuthProvider>
         <ThemeProvider>
-          <GameProvider>
-            <PageRouter />
-            <GlobalStyles />
-          </GameProvider>
+          <PageRouter />
+          <GlobalStyles />
         </ThemeProvider>
       </AuthProvider>
     </React.StrictMode>
